refactor: change composable name
