+++ apps/web/composables/useFooter/index.ts
@@ -1,0 +1,1 @@
+export * from './useFooter';
