+++ apps/web/composables/useFooter/useFooter.ts
