--- apps/web/composables/useFooterSettings/index.ts
@@ -1,1 +1,0 @@
-export * from './useFooterSettings';
